% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetops.R
\name{as_pixels}
\alias{as_pixels}
\title{Convert subpixel tops to integer pixel positions}
\usage{
as_pixels(tops)
}
\arguments{
\item{tops}{Two-column matrix of (row, col) positions.}
}
\value{
An integer two-column matrix suitable for \code{\link[=grow_crowns]{grow_crowns()}}.
}
\description{
Convert subpixel tops to integer pixel positions
}
